moved integration test to integration dir
